+++ engine/language-server/src/main/scala/org/enso/languageserver/protocol/json/JsonConnectionController.scala
@@ -505,7 +505,8 @@ class JsonConnectionController(
         .props(requestTimeout, localLibraryManager),
       LibraryGetMetadata -> LibraryGetMetadataHandler.props(),
       LibraryPreinstall  -> LibraryPreinstallHandler.props(),
-      LibraryPublish     -> LibraryPublishHandler.props(),
+      LibraryPublish -> LibraryPublishHandler
+        .props(requestTimeout, localLibraryManager),
       LibrarySetMetadata -> LibrarySetMetadataHandler.props()
     )
   }