+++ engine/launcher/src/main/scala/org/enso/launcher/Launcher.scala
@@ -343,6 +343,52 @@ case class Launcher(cliOptions: GlobalCLIOptions) {
     0
   }
 
+  /** Uploads a library to a repository.
+    *
+    * @param path path to the library, if not specified, the current working
+    *             directory and its ancestors are searched for an Enso project
+    *             to upload
+    * @param uploadUrl a URL of an upload endpoint of a repository; if not
+    *                  specified, falls back to the default Enso repository
+    * @param authToken a token to use for authentication
+    * @param logLevel log level for the language server
+    * @param useSystemJVM if set, forces to use the default configured JVM,
+    *                     instead of the JVM associated with the engine version
+    * @param jvmOpts additional options to pass to the launched JVM
+    * @param additionalArguments additional arguments to pass to the runner
+    * @return exit code of the launched program
+    */
+  def uploadLibrary(
+    path: Option[Path],
+    uploadUrl: Option[String],
+    authToken: Option[String],
+    logLevel: LogLevel,
+    useSystemJVM: Boolean,
+    jvmOpts: Seq[(String, String)],
+    additionalArguments: Seq[String]
+  ): Int = {
+    val settings = runner
+      .uploadLibrary(
+        path,
+        uploadUrl.getOrElse {
+          throw new IllegalArgumentException(
+            "The default repository is currently not defined. " +
+            "You need to explicitly specify the `--upload-url`."
+          )
+        },
+        authToken.orElse(LauncherEnvironment.getEnvVar("ENSO_AUTH_TOKEN")),
+        cliOptions.hideProgress,
+        logLevel,
+        cliOptions.internalOptions.logMasking,
+        additionalArguments
+      )
+      .get
+
+    runner.withCommand(settings, JVMSettings(useSystemJVM, jvmOpts)) {
+      command => command.run().get
+    }
+  }
+
   /** Prints the value of `key` from the global configuration.
     *
     * If the `key` is not set in the config, sets exit code to 1 and prints a