+++ lib/scala/downloader/src/main/scala/org/enso/downloader/http/HTTPDownload.scala
@@ -157,6 +157,7 @@ object HTTPDownload {
         "akka.library-extensions",
         ConfigValueFactory.fromAnyRef(Seq.empty.asJava)
       )
+      .withValue("akka.daemonic", ConfigValueFactory.fromAnyRef("on"))
       .withValue("akka.loggers", ConfigValueFactory.fromAnyRef(loggers))
       .withValue(
         "akka.logging-filter",